Move parser interface to new package
